Refactor of components file naming
