+++ webapp/app/components/index.js
@@ -1,5 +1,5 @@
-export { Message } from './Message/Message.component';
-export { MessageBox } from './MessageBox/MessageBox.component';
-export { MessageList } from './MessageList/MessageList.component';
+export { Message } from './message/message.component';
+export { MessageBox } from './messageBox/messageBox.component';
+export { MessageList } from './messageList/messageList.component';
 export { RoomList } from './roomList/roomList.component';
-export { UserAvatar } from './UserAvatar/UserAvatar.component';
+export { UserAvatar } from './userAvatar/userAvatar.component';
