+++ webapp/app/components/userAvatar/userAvatar.component.js
